requirements.txt
asgiref==3.3.4
attrs==21.2.0
Django==3.2.3
django-autoslug==1.9.8
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-js-asset==1.2.2
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pdfkit==0.6.0
Pillow==8.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
